Styled pages/forms, fixed edit_recipe whitespace issue (ref: StackOverflow), added user profile image & updated home to feature 4 random recipes.
also fix bug in my_recipe by  consolidating the calls to users_coll.find_one() into a single call, fixed the call to count_documents() by calling it directly on the recipes_coll collection and passing the relevant filter. Added a check to ensure that the user document exists before trying to access its properties. If the user is not found, it returns a 404 error with a "User not found!" message. This prevents potential NoneType errors if the user does not exist in the database.